requirements.txt
icarus
django>=2.1
djangorestframework>=3.9.1
